@@ -1,19 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
   <ItemGroup>
     <Compile Remove="BitSet.cs" />
     <Compile Remove="ResizableBitSet.cs" />
   </ItemGroup>
-
   <ItemGroup>
     <None Remove="CharClass.cs.old" />
   </ItemGroup>
-
 </Project>
