@@ -6,23 +6,18 @@
  */
 namespace NRegEx;
 
-public class Match : Group
+public class Match(Regex Source, string Input, bool Success, string Name = "", int InclusiveStart = 0, int ExclusiveEnd = int.MaxValue, string? Value = null) : Group(Success, Name, InclusiveStart, ExclusiveEnd, Value)
 {
-    public readonly string Input;
-    public readonly Regex Source;
-    public readonly List<Group> Groups = new();
+    public readonly string Input = Input;
+    public readonly Regex Source = Source;
+    public readonly List<Group> Groups = [];
     public new virtual Group this[string name]
         => this.Groups.FirstOrDefault(g => g.Name == name) ?? Empty;
     public new virtual Group this[int index] 
         => this.Groups[index];
     public new virtual int Count
         => this.Groups.Count;
-    public Match(Regex Source, string Input, bool Success, string Name = "", int InclusiveStart = 0, int ExclusiveEnd = int.MaxValue, string? Value = null)
-        : base(Success, Name, InclusiveStart, ExclusiveEnd, Value)
-    {
-        this.Source = Source;
-        this.Input = Input;
-    }
+
     public override string ToString()
         => this.Value ?? base.ToString();
 }
